Add some feelings we could possibly draw from
We want to convey the feeling of using Rust at its best to peole who haven't worked with it, yet. Here are a few I came up with. My hope is that we might be able to gather then feelings, find the tones, and then use those to create pithy expressions.